requirements.txt
faiss-cpu
sentence-transformers
spacy
openai
tqdm
streamlit
nltk
requests
beautifulsoup4
pandas
